djangox/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
crispy-bootstrap5==0.7
cryptography==39.0.1
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.52.0
django-crispy-forms==2.0
django_debug_toolbar==3.8.1
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.2
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
urllib3==1.26.14
whitenoise==6.4.0
plotly==5.14.1
numpy==1.24.2
pandas==2.0.0
sympy==1.11.1
